CORE-8728 Add contract verification to the beginning of the Finality flow too
