+++ components/ledger/ledger-utxo-flow/src/main/kotlin/net/corda/ledger/utxo/flow/impl/flows/finality/UtxoFinalityBase.kt
@@ -3,6 +3,8 @@ package net.corda.ledger.utxo.flow.impl.flows.finality
 import net.corda.ledger.common.flow.transaction.TransactionSignatureService
 import net.corda.ledger.utxo.flow.impl.persistence.UtxoLedgerPersistenceService
 import net.corda.ledger.utxo.flow.impl.transaction.UtxoSignedTransactionInternal
+import net.corda.ledger.utxo.flow.impl.transaction.verifier.UtxoLedgerTransactionVerifier
+import net.corda.ledger.utxo.flow.impl.transaction.verifier.UtxoTransactionMetadataVerifier
 import net.corda.sandbox.CordaSystemFlow
 import net.corda.v5.application.crypto.DigitalSignatureAndMetadata
 import net.corda.v5.application.flows.CordaInject
@@ -56,6 +58,14 @@ abstract class UtxoFinalityBase : SubFlow<UtxoSignedTransaction> {
     ):UtxoSignedTransactionInternal {
         verifySignature(transaction.id, signature)
         return transaction.addSignature(signature)
+    }
+
+    protected fun verifyTransaction(signedTransaction: UtxoSignedTransaction) {
+        UtxoTransactionMetadataVerifier(signedTransaction.metadata).verify()
+        val ledgerTransactionToCheck = signedTransaction.toLedgerTransaction()
+        val verifier = UtxoLedgerTransactionVerifier(ledgerTransactionToCheck)
+        verifier.verifyPlatformChecks(signedTransaction.notary)
+        verifier.verifyContracts()
     }
 
 }
